requirements.txt
fastapi==0.110.2
stripe==9.5.0
pymongo==4.7.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
uvicorn==0.29.0
pydantic==2.7.1
passlib==1.7.4
bcrypt==4.1.2
fastapi-utilities==0.2.0
